Update and rename desEncrypt.java to DesEncryptUtils.java
